src/photo_to_3d.py
#!/usr/bin/env python
# -*- coding: utf-8 -*-

"""
Module for converting photos to 3D models.
"""

import os
import numpy as np
from PIL import Image
import trimesh
import logging

# Set up logging
logger = logging.getLogger(__name__)


def photo_to_model(input_path, output_path, resolution=128, config=None):
    """
    Convert a photo to a 3D model.
    
    Args:
        input_path (str): Path to the input photo.
        output_path (str): Path to save the output 3D model.
        resolution (int): Resolution of the output 3D model.
        config (dict): Configuration parameters.
    
    Returns:
        bool: True if successful, False otherwise.
    """
    try:
        logger.info(f"Processing image: {input_path}")
        
        # Load the image
        image = Image.open(input_path)
        
        # Resize image if needed
        if max(image.size) > 1024:
            image.thumbnail((1024, 1024))
        
        # Convert to grayscale for depth estimation
        gray_image = image.convert('L')
        depth_map = np.array(gray_image)
        
        # Normalize depth map
        depth_map = depth_map / 255.0
        
        # Create a simple 3D mesh from the depth map
        # In a real implementation, this would use a more sophisticated
        # depth estimation algorithm or neural network
        
        # Create a grid of vertices
        x = np.linspace(0, 1, resolution)
        y = np.linspace(0, 1, resolution)
        xv, yv = np.meshgrid(x, y)
        
        # Create vertices
        vertices = np.zeros((resolution, resolution, 3))
        vertices[:, :, 0] = xv
        vertices[:, :, 1] = yv
        vertices[:, :, 2] = depth_map[
            np.linspace(0, depth_map.shape[0]-1, resolution).astype(int),
            np.linspace(0, depth_map.shape[1]-1, resolution).astype(int)
        ]
        
        # Reshape vertices to a list
        vertices = vertices.reshape(-1, 3)
        
        # Create faces (triangles)
        faces = []
        for i in range(resolution-1):
            for j in range(resolution-1):
                v0 = i * resolution + j
                v1 = i * resolution + j + 1
                v2 = (i + 1) * resolution + j
                v3 = (i + 1) * resolution + j + 1
                
                # Add two triangles for each grid cell
                faces.append([v0, v1, v2])
                faces.append([v1, v3, v2])
        
        faces = np.array(faces)
        
        # Create a mesh
        mesh = trimesh.Trimesh(vertices=vertices, faces=faces)
        
        # Determine the output format
        if output_path.lower().endswith('.stl'):
            mesh.export(output_path, file_type='stl')
        else:  # Default to OBJ
            mesh.export(output_path, file_type='obj')
        
        logger.info(f"3D model saved to: {output_path}")
        return True
        
    except Exception as e:
        logger.error(f"Error converting photo to 3D model: {str(e)}")
        return False
